refactor(preview): split ng worker rendering helpers
